src/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0
pyserial>=3.5
scipy>=1.7.0
